Bug in the ConstructorArguments from CustomAttributes for null array of enum
This has to do with the structure of the blob where the custom attribute is specified.
Array values start with an integer indicating the number of elements in the array, then the item values concatentated together.
A null array is represented using a length of -1.
An enum argument is represented using the byte 0x55 followed by a string specifying the name and assembly of the enum type.
Unfortunately, what happens if you pass an array of enum as null is that enum name is lost

Fix #7346